# Optimize SeriesGrouper & aggregators.merge