--- a/plugins/aggregators/merge/merge.go
+++ b/plugins/aggregators/merge/merge.go
@@ -36,13 +36,7 @@ func (a *Merge) SampleConfig() string {
 }
 
 func (a *Merge) Add(m telegraf.Metric) {
-	tags := m.Tags()
-	for _, field := range m.FieldList() {
-		err := a.grouper.Add(m.Name(), tags, m.Time(), field.Key, field.Value)
-		if err != nil {
-			a.log.Errorf("Error adding metric: %v", err)
-		}
-	}
+	a.grouper.AddMetric(m)
 }
 
 func (a *Merge) Push(acc telegraf.Accumulator) {