--- a/plugins/aggregators/merge/merge_test.go
+++ b/plugins/aggregators/merge/merge_test.go
@@ -4,9 +4,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/influxdata/telegraf"
-	"github.com/influxdata/telegraf/testutil"
 	"github.com/stretchr/testify/require"
+
+	"github.com/influxdata/telegraf"
+	"github.com/influxdata/telegraf/metric"
+	"github.com/influxdata/telegraf/testutil"
 )
 
 func TestSimple(t *testing.T) {
@@ -183,4 +185,69 @@ func TestReset(t *testing.T) {
 	}
 
 	testutil.RequireMetricsEqual(t, expected, acc.GetTelegrafMetrics())
+}
+
+var m1, _ = metric.New(
+	"mymetric",
+	map[string]string{
+		"host":        "host.example.com",
+		"mykey":       "myvalue",
+		"another key": "another value",
+	},
+	map[string]interface{}{
+		"f1": 1,
+		"f2": 2,
+		"f3": 3,
+		"f4": 4,
+		"f5": 5,
+		"f6": 6,
+		"f7": 7,
+		"f8": 8,
+	},
+	time.Now(),
+)
+var m2, _ = metric.New(
+	"mymetric",
+	map[string]string{
+		"host":        "host.example.com",
+		"mykey":       "myvalue",
+		"another key": "another value",
+	},
+	map[string]interface{}{
+		"f8":  8,
+		"f9":  9,
+		"f10": 10,
+		"f11": 11,
+		"f12": 12,
+		"f13": 13,
+		"f14": 14,
+		"f15": 15,
+		"f16": 16,
+	},
+	m1.Time(),
+)
+
+func BenchmarkMergeOne(b *testing.B) {
+	var merger Merge
+	merger.Init()
+	var acc testutil.NopAccumulator
+
+	for n := 0; n < b.N; n++ {
+		merger.Reset()
+		merger.Add(m1)
+		merger.Push(&acc)
+	}
+}
+
+func BenchmarkMergeTwo(b *testing.B) {
+	var merger Merge
+	merger.Init()
+	var acc testutil.NopAccumulator
+
+	for n := 0; n < b.N; n++ {
+		merger.Reset()
+		merger.Add(m1)
+		merger.Add(m2)
+		merger.Push(&acc)
+	}
 }