@phemmer 2020-11-11
I haven't a clue whats up with the test failure.

macdeps:

```
github.com/influxdata/telegraf/metric imports
	hash/maphash: malformed module path "hash/maphash": missing dot in first path element
```

... No it's not