@reimda 2020-11-11
I took a look at the macdeps failure but I haven't figured out yet why it's failing.  First I reran it and it failed again.  I noticed when rerunning it that macdeps installs go version 1.13.6 when it intends to install the latest go package from homebrew.  Maybe 'go mod tidy' fails only on go 1.13.6?  (The other CI steps use 1.14 or 1.15.)