@phemmer 2020-11-11
Ah, that would explain it. maphash was introduced in 1.14.

Given that `go.mod` has the version set to 1.15, should we not even be testing against lower versions?