@reimda 2020-11-11
I just noticed that hash/maphash is new in 1.14 too.
https://golang.org/doc/go1.14#hash/maphash

Telegraf is currently supporting go 1.14 and 1.15 even though go.mod has the version set to 1.15.  Fortunately the 1.14 CI steps are working even with the go.mod version set to 1.15.

To fix this failure we'll have to fix the macdeps job to use a newer golang version.  I think it's just a matter of updating .circleci/config.yml to run 'brew update' before running 'brew install go'.  Could you make that change and try it out in this PR?