@reimda 2020-11-11
It looks like that worked.  Unfortunately 'brew update' took over four minutes to run.  We may need to look for a faster way to get an up to date golang install, but you don't need to worry about that for this PR.  Thanks for your help debugging CI!