@phemmer 2020-11-11
Yay.

> We may need to look for a faster way to get an up to date golang install

Maybe it should use the official package/archive. I know I've seen issues on the golang issue tracker where they won't support people using the homebrew package, and tell these people to go reproduce issues using the official package. I'm unfamiliar with the background, but some googling suggests that homebrew does janky things with the package.