@reimda 2020-12-04
I updated the golang version for mac ci in #8516.  Could you remove the 'brew update' from .circleci/config.yml and rebase this PR?  Thanks!